수정: Character.js 내부 테이블들 컴포넌트로 분리
